organize app-api dir -- what do you think?
